fix(storage): harden proof proto conversions
Use semantic storage proto wrappers for persisted proof and certificate fields
and make SP1 proof serialization validate version strings at the concrete
encode/decode boundary.

Keep the core proof envelope generic while storing aggchain and pessimistic
proof public values in explicit domain wrappers.
