--- proto/agglayer/storage/v0/storage_proof.proto
@@ -1,53 +1,0 @@
-syntax = "proto3";
-
-package agglayer.storage.v0;
-
-// Proof system that produced the opaque payload bytes.
-enum ProofSystem {
-  // Default value, should not be used.
-  PROOF_SYSTEM_UNSPECIFIED = 0;
-  // SP1 proof system.
-  PROOF_SYSTEM_SP1 = 1;
-}
-
-// Mode of the SP1 proof.
-enum ProofMode {
-  // Default value, should not be used.
-  PROOF_MODE_UNSPECIFIED = 0;
-  // Compressed (STARK) SP1 proof.
-  PROOF_MODE_COMPRESSED = 1;
-  // Plonk SNARK wrapper over an SP1 proof.
-  PROOF_MODE_PLONK = 2;
-  // Groth16 SNARK wrapper over an SP1 proof.
-  PROOF_MODE_GROTH16 = 3;
-}
-
-// Storage envelope for a zero-knowledge proof.
-//
-// The protobuf envelope is stable and self-contained.
-// The payload bytes are opaque to protobuf, but their encoding is selected by
-// `proof_system`, `mode`, and `version`.
-// For SP1 proofs, the payload fields currently store version-specific
-// Rust-serialized proof-system data rather than a generic cross-SDK format.
-message Proof {
-  // Which proof system produced this proof.
-  ProofSystem proof_system = 1;
-
-  // Preserved end to end and used to select the payload deserializer.
-  string version = 2;
-
-  // Proof mode inside the selected proof system.
-  ProofMode mode = 3;
-
-  // Opaque serialized proof payload.
-  // Decoding requires `proof_system`, `mode`, and `version`.
-  bytes proof = 4;
-
-  // Opaque serialized public values.
-  // Decoding requires `proof_system`, `mode`, and `version`.
-  bytes public_values = 5;
-
-  // Opaque serialized verifying key.
-  // Decoding requires `proof_system`, `mode`, and `version`.
-  bytes vkey = 6;
-}
